Remove launder sites the ROM does not need (5/5)
Chunk 5 of 5 of the delaunder sweep, split so each part lands under the
PR validator's 200-file cap and every file gets a real CI byte gate.

174 files. Each site was removed only when the ROM still agreed with the
result, one site at a time, by tools/delaunder.py.

+++ src/func_ov062_021164e8.cpp
@@ -35,7 +35,7 @@ do_block:
     {
         void *m = *(void **)(c + 0x3f8);
         *(u16 *)(c + 0x94) = *(s16 *)((char *)m + 0x8e);
-        *(int *)((long long)(int)(c + 0x128)) |= 2;
+        *(int *)(c + 0x128) |= 2;
         Chuckya_ChangeState(c, &data_ov062_0211dec0);
         if (*(int *)((char *)*(void **)(c + 0x3f8) + 8) == 2) {
             *(int *)(c + 0xa8) = 0x50000;
