+++ src/func_ov062_0211a740.cpp
@@ -29,7 +29,7 @@ extern "C" void func_ov062_0211a740(char* c)
     case 0:
         if (_Z14ApproachLinearRsss((s16*)(c + 0x94), *(s16*)(c + 0x3a8), 0x800) != 0) {
             _ZN9ModelAnim7SetAnimEP8BCA_Filei5Fix12IiEj((void*)(c + 0x300), (void*)data_ov062_0211e03c[1], 0, 0x1000, 0);
-            *(u8*)((((long long)(int)(c + 0x390)))) += 1;
+            *(u8*)(c + 0x390) += 1;
             func_0201267c(0xec, c + 0x74);
             if (*(int*)(*(int*)(c + 0x398) + 8) == 0) {
                 func_02012790(0xa);
@@ -61,7 +61,7 @@ extern "C" void func_ov062_0211a740(char* c)
                 v.z = z;
             }
             if (_ZN6Player11ShowMessageER9ActorBasejPK7Vector3hh(*(void**)(c + 0x398), c, msg, &v, 1, 0) != 0)
-                *(u8*)((((long long)(int)(c + 0x390)))) += 1;
+                *(u8*)(c + 0x390) += 1;
         }
         return;
     case 2:
