+++ src/func_ov084_0212ddbc.c
@@ -87,8 +87,8 @@ cold:
     if (*(s32 *)(c + 0x204) <= (*(s32 *)(c + 0x210) >> 1))
         return;
     {
-        u32 *p18c = (u32 *)(((long long)(int)(c + 0x18c)));
-        u32 *pb0 = (u32 *)(((long long)(int)(c + 0xb0)));
+        u32 *p18c = (u32 *)(c + 0x18c);
+        u32 *pb0 = (u32 *)(c + 0xb0);
         *p18c &= ~1u;
         *pb0 |= 0x10000000u;
     }
