+++ src/func_ov098_021385e0.c
@@ -1,6 +1,6 @@
 extern short data_02082214[];
 
-#define ADDR(p) ((int *)(((long long)(int)(p))))
+#define ADDR(p) ((int *)(p))
 
 void func_ov098_021385e0(char *self)
 {
