+++ src/func_ov102_0214baa0.c
@@ -5,9 +5,9 @@ extern void func_0203568c(int *value, int target);
 extern void func_02035684(int *value, int target);
 
 #define LAUNDER_PTR(ptr) \
-    ((int *)(int)(((long long)(int)(ptr))))
+    ((int *)(int)(ptr))
 #define LAUNDER_32(base, offset) \
-    ((int *)(int)(((long long)((int)(base) + (offset)))))
+    ((int *)(int)((long long)((int)(base) + (offset))))
 
 void func_ov102_0214baa0(char *self)
 {
