+++ src/unnamed/ov063/func_ov063_0211934c.cpp
@@ -34,7 +34,7 @@ extern "C" void func_ov063_0211934c(char *c)
             *(s16 *)(c + 0x92) = *(s16 *)(c + 0x570);
             *(s16 *)(c + 0x94) = *(s16 *)(c + 0x572);
             *(s16 *)(c + 0x96) = *(s16 *)(c + 0x574);
-            s16 *src = (s16 *)(((int)c + 0x92));
+            s16 *src = (s16 *)((int)c + 0x92);
             *(s16 *)(c + 0x8c) = src[0];
             *(s16 *)(c + 0x8e) = src[1];
             *(s16 *)(c + 0x90) = src[2];
@@ -43,7 +43,7 @@ extern "C" void func_ov063_0211934c(char *c)
             if (cap == 0)
                 return;
             {
-                u16 *p = (u16 *)(((int)cap + 0x5d4));
+                u16 *p = (u16 *)((int)cap + 0x5d4);
                 *p &= ~2;
             }
         }
@@ -56,7 +56,7 @@ extern "C" void func_ov063_0211934c(char *c)
         if (_ZN6Player9StartTalkER9ActorBaseb(r4, c, 1) == 0)
             return;
         {
-            u8 *q = (u8 *)(((long long)(int)(c + 0x5ce)));
+            u8 *q = (u8 *)(c + 0x5ce);
             *q = *q + 1;
         }
         return;
@@ -67,7 +67,7 @@ extern "C" void func_ov063_0211934c(char *c)
         *(void **)(c + 0x48c) = _ZN5Actor10FindWithIDEj(*(u32 *)(c + 0x490));
         found = *(void **)(c + 0x48c);
         if (found != 0) {
-            s32 *cnt = (s32 *)(((long long)(int)((char *)found + 0x180)));
+            s32 *cnt = (s32 *)((char *)found + 0x180);
             *cnt = *cnt + 1;
         }
         found = *(void **)(c + 0x48c);
@@ -78,7 +78,7 @@ extern "C" void func_ov063_0211934c(char *c)
         }
         *(void **)(c + 0x48c) = 0;
         {
-            u8 *q = (u8 *)(((long long)(int)(c + 0x5ce)));
+            u8 *q = (u8 *)(c + 0x5ce);
             *q = *q + 1;
         }
         func_0201267c(0xf8, c + 0x74);
@@ -96,13 +96,13 @@ extern "C" void func_ov063_0211934c(char *c)
             if (found != 0 && *(s32 *)((char *)found + 0x180) == 5) {
                 func_ov063_02116244((char *)found);
                 {
-                    u16 *p = (u16 *)(((long long)(int)(c + 0x5c6)));
+                    u16 *p = (u16 *)(c + 0x5c6);
                     *p = *p + 1;
                 }
             }
         }
         {
-            u8 *q = (u8 *)(((long long)(int)(c + 0x5ce)));
+            u8 *q = (u8 *)(c + 0x5ce);
             *q = *q + 1;
         }
         return;
@@ -110,7 +110,7 @@ extern "C" void func_ov063_0211934c(char *c)
     case 3:
         if (*(u16 *)(c + 0x500 + 0xc6) != 0) {
             if (_ZN5Sound15PlaySecretSoundEP5ActorPt(
-                    c, (u16 *)(((long long)(int)(c + 0x5c6)))) == 0)
+                    c, (u16 *)(c + 0x5c6)) == 0)
                 return;
             _ZN9ActorBase18MarkForDestructionEv(c);
             if (((*(u8 *)(c + 0x113)) & 0xf) >= 6)
@@ -125,7 +125,7 @@ extern "C" void func_ov063_0211934c(char *c)
             *(s16 *)(c + 0x92) = *(s16 *)(c + 0x570);
             *(s16 *)(c + 0x94) = *(s16 *)(c + 0x572);
             *(s16 *)(c + 0x96) = *(s16 *)(c + 0x574);
-            s16 *src = (s16 *)(((int)c + 0x92));
+            s16 *src = (s16 *)((int)c + 0x92);
             *(s16 *)(c + 0x8c) = src[0];
             *(s16 *)(c + 0x8e) = src[1];
             *(s16 *)(c + 0x90) = src[2];
@@ -134,7 +134,7 @@ extern "C" void func_ov063_0211934c(char *c)
             if (cap2 == 0)
                 return;
             {
-                u16 *p = (u16 *)(((int)cap2 + 0x5d4));
+                u16 *p = (u16 *)((int)cap2 + 0x5d4);
                 *p &= ~2;
             }
         }
@@ -153,14 +153,14 @@ extern "C" void func_ov063_0211934c(char *c)
             *(s16 *)(c + 0x92) = *(s16 *)(c + 0x570);
             *(s16 *)(c + 0x94) = *(s16 *)(c + 0x572);
             *(s16 *)(c + 0x96) = *(s16 *)(c + 0x574);
-            s16 *src = (s16 *)(((int)c + 0x92));
+            s16 *src = (s16 *)((int)c + 0x92);
             *(s16 *)(c + 0x8c) = src[0];
             *(s16 *)(c + 0x8e) = src[1];
             *(s16 *)(c + 0x90) = src[2];
             
             cap3 = _ZN8CapEnemy15RespawnIfHasCapEv(c);
             if (cap3 != 0) {
-                u16 *p = (u16 *)(((int)cap3 + 0x5d4));
+                u16 *p = (u16 *)((int)cap3 + 0x5d4);
                 *p &= ~2;
             }
         }
